3rd Commit
I added player movement and third person camera. I scripted the player move on key press, player turn on mouse location and the smoothness in the player model rotation. I also implemented cinemachine camera to make the aim camera and the normal camera that follows the player in third person.