4th Commit
I finished the movement . I finalized the movement of the player by adding the jumping and the sprinting mechanic. I had a small problem with the ground and the player where the player model sinks a bit into the ground. I fixed it by making the player controller and the body the same height and I added a layer to the ground.