6th Commit
In this commit I implemented a zoom in on click as well as the crosshair. When the user right clicks, the camera zooms in and the crosshair range lessens in size for more accuracy.  I also implemented a damage system. It is only on testing right now since I put the code on a truck and did a debug on it to see if the object takes damage. I also locked the mouse cursor so it wouldnt divert the attention of the player mid game. Lastly, I added some assets from the unity store (War FX) and gave them some use. There is the muzzle flash: when the user presses left click, the particle shows for a brief second and the bullet impact on hit. With that I made the gun an automatic rifle and gave it fire rate.