7th Commit
Today I fixed the muzzle fire effect to be on the gun tip at all times. I implemented the reloading mechanic on the gun and the animations to come with it. I have a small problem with the reload while the player is aiming and walking and will try to fix that.  Lastly, I added the enemy model and gave it colliders and a nav mesh. I will start working on the AI.