9th Commit
Today I made the enemy AI to follow the player. I used the radius as a behaviour to the enemy so when the player is spotted in the visionRadius of the enemy he walks toward him and when the enemy gets to the shootingRadius he starts shooting. I added damage to both the player and the enemy. When the enemy is shot down to 0HP he will stay dead for 5 seconds and be destroyed. For now there is no animation on enemy death but it will be implemented. The player is instantly destroyed when he hits 0 but there is no animation for now either. I added blood particles to the bullets that hit the enemy as well. To know which object got hit by the enemy and player bullets I added a debug.log in the script. I fixed a bug that when I shoot outside the enemy visionRadius the enemy takes damage but when im inside the radius he doesnt. So i removed the enemies' rigid body and that worked just fine.