10th Commit
Today I made an animator controller for the enemy and gave him animation on all his reactions.. I also gave the enemy fire animation a muzzle flash particle effect like the player. Lastly I added an enemy drone model and for the next commit, the script and the animations will be ready on the drone enemy.