12th Commit
Today I added animations to the EnemyDrone when it shoots and dies. I added muzzle flash and the flamethrower weapons on it which gave me quite a bit of problems. The flamethrower gave problems on when the particles start appearing from the drone's muzzle. The gun fire from the gun was ok but the flamethrower was starting after 7 seconds. I tried everything with the start lifetime and the start speed but it still didnt work. So I searched for online help and got the solution. I needed to tick the loop, prewarm boxes on the particle system and select scaling mode as hierarchy. As for being hit by player bullets I changed the effect from blood to metal sparks since robots dont bleed. Lastly I added a destroy effect on the drone so that when the drone's HP reaches zero an explosion particle goes off.