13th Commit
Today I added the health bar on the enemy and the player. I added a slider on the health bar so that when either the enemy or player take damage the slider will go down depending on the damage. I also made the enemy health bar as a child to the enemy to follow the enemy. I encountered a problem that when I look at the enemy at a certain angle the health bar disappears or looks smaller. I fixed the problem by making a "RotateHealthbarUI" script and attaching the main camera onto the script which was attached onto the enemy canvas. This all applies to the drone as well.