14th Commit
In this commit I added the reload UI with bullet and magazine amount. I decided to add a magazine mechanic to make it feel a bit different than the average shooter reload mechanic. When the player runs out of ammo and magazines, a text will show up each time the player tries to shoot, telling him that he is out of ammo. I also made a pause and end game screen with functioning buttons. The end game screen is basically the death screen and I added another camera to hover over the map displaying the menu and showing the area where the player died to make it seem a bit more interesting.