16th Commit
In this commit I did the gate system controller scripts and the door animation. I made a key game object to open the gate. I also made the cursor change colour when it hovers on an interactable object. What I managed to do was the highlight cursor on interactables, pick up the key by presing 'F' and store it within the player but I wasnt able to open the gate with the key as im getting an error with the animation part of the code. For now it will be commented out until I fix it.