18th Commit
In this commit i added an objective system. I listed the objectives on a canvas and when the player finishes 1 objective, the text changes to green to indicate that the task is finished. I also added an extraction area. The player spawns near an armored van and when he finishes his mission he can just simply press 'F' and it will display the menu.